{{cookiecutter.project_name}}/requirements/test.txt
-r base.txt
flake8==3.0.4
isort==4.2.5
pytest-aiohttp==0.1.3
